backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pandas==2.2.3
numpy==2.2.1
scipy==1.15.1
matplotlib==3.10.0
seaborn==0.13.2
reportlab==4.2.5
openpyxl==3.1.5
xlrd==2.0.1
odfpy==1.4.1
python-multipart==0.0.20
python-dotenv==1.0.1
jinja2==3.1.5
pytest==8.3.4
httpx==0.28.1
